--- a/lib/quick-sort.js
+++ b/lib/quick-sort.js
@@ -15,6 +15,8 @@
 // sorting in C++. By using our own JS-implemented Quick Sort (below), we get
 // a ~3500ms mean speed-up in `bench/bench.html`.
 
+function SortTemplate(comparator) {
+
 /**
  * Swap the elements indexed by `x` and `y` in the array `ary`.
  *
@@ -101,6 +103,15 @@ function doQuickSort(ary, comparator, p, r) {
   }
 }
 
+  return doQuickSort;
+}
+
+function cloneSort(comparator) {
+  let template = SortTemplate.toString();
+  let templateFn = new Function(`return ${template}`)();
+  return templateFn(comparator);  // Invoke template to get doQuickSort
+}
+
 /**
  * Sort the given array in-place with the given comparator function.
  *
@@ -109,6 +120,13 @@ function doQuickSort(ary, comparator, p, r) {
  * @param {function} comparator
  *        Function to use to compare two items.
  */
+
+let sortCache = new WeakMap();  // Cache for specialized sorts.
 exports.quickSort = function (ary, comparator) {
+  let doQuickSort = sortCache.get(comparator);
+  if (doQuickSort === void 0) {
+    doQuickSort = cloneSort(comparator);
+    sortCache.set(comparator, doQuickSort);
+  }
   doQuickSort(ary, comparator, 0, ary.length - 1);
 };