--- a/lib/quick-sort.js
+++ b/lib/quick-sort.js
@@ -109,7 +109,7 @@ function doQuickSort(ary, comparator, p, r) {
 function cloneSort(comparator) {
   let template = SortTemplate.toString();
   let templateFn = new Function(`return ${template}`)();
-  return templateFn(comparator);  // Invoke template to get doQuickSort
+  return templateFn(comparator);
 }
 
 /**
@@ -121,7 +121,7 @@ function cloneSort(comparator) {
  *        Function to use to compare two items.
  */
 
-let sortCache = new WeakMap();  // Cache for specialized sorts.
+let sortCache = new WeakMap();
 exports.quickSort = function (ary, comparator, start = 0) {
   let doQuickSort = sortCache.get(comparator);
   if (doQuickSort === void 0) {