--- a/lib/source-map-consumer.js
+++ b/lib/source-map-consumer.js
@@ -578,8 +578,6 @@ BasicSourceMapConsumer.prototype._parseMappings =
 
         generatedMappings.push(mapping);
         if (typeof mapping.originalLine === 'number') {
-          // This code used to just do: originalMappings.push(mapping).
-          // Now it sorts original mappings already by source during parsing.
           let currentSource = mapping.source;
           while (originalMappings.length <= currentSource) {
             originalMappings.push(null);
@@ -595,8 +593,6 @@ BasicSourceMapConsumer.prototype._parseMappings =
     sortGenerated(generatedMappings, subarrayStart);
     this.__generatedMappings = generatedMappings;
 
-    // The code used to sort the whole array:
-    // quickSort(originalMappings, util.compareByOriginalPositions);
     for (var i = 0; i < originalMappings.length; i++) {
       if (originalMappings[i] != null) {
         quickSort(originalMappings[i], util.compareByOriginalPositionsNoSource);