--- a/lib/util.js
+++ b/lib/util.js
@@ -385,6 +385,38 @@ function compareByOriginalPositions(mappingA, mappingB, onlyCompareOriginal) {
 }
 exports.compareByOriginalPositions = compareByOriginalPositions;
 
+function compareByOriginalPositionsNoSource(mappingA, mappingB, onlyCompareOriginal) {
+  // var cmp = strcmp(mappingA.source, mappingB.source);
+  // if (cmp !== 0) {
+  //   return cmp;
+  // }
+
+  var cmp
+
+  cmp = mappingA.originalLine - mappingB.originalLine;
+  if (cmp !== 0) {
+    return cmp;
+  }
+
+  cmp = mappingA.originalColumn - mappingB.originalColumn;
+  if (cmp !== 0 || onlyCompareOriginal) {
+    return cmp;
+  }
+
+  cmp = mappingA.generatedColumn - mappingB.generatedColumn;
+  if (cmp !== 0) {
+    return cmp;
+  }
+
+  cmp = mappingA.generatedLine - mappingB.generatedLine;
+  if (cmp !== 0) {
+    return cmp;
+  }
+
+  return strcmp(mappingA.name, mappingB.name);
+}
+exports.compareByOriginalPositionsNoSource = compareByOriginalPositionsNoSource;
+
 /**
  * Comparator between two mappings with deflated source and name indices where
  * the generated positions are compared.