@7rulnik 2021-01-19
@benthemonkey I will close it and will publish it as my own fork to have more flexibility on review and releases. I will add credits to you because you wrote the issue in the original repo and took my attention to the problem with this library. Hope you don't mind!