@7rulnik 2021-02-07
@benthemonkey check https://www.npmjs.com/package/source-map-js :)